+++ src/app/demo/chatapp/user/[slug]/login/page.tsx
@@ -1,5 +1,7 @@
 "use client";
 
+
+
 import 'bootstrap/dist/css/bootstrap.min.css';
 import { NavBar, HeadingBar, CustomBody, SpaceBlock, Footer } from "front-end-component-kit";
 import { Container, Row, Col, Button, Form } from "react-bootstrap";
@@ -14,90 +16,52 @@ type Message = {
   message: string;
 };
 
-export default function ChatClient() {
-  const [messages, setMessages] = useState<Message[]>([]);
-  const [input, setInput] = useState("");
-  const wsRef = useRef<WebSocket | null>(null);
 
-  const chatHash = "6LRcGlCjvNB"; // hardcoded for now
-  const sender = "Abhi";
-  const receiver = "Anny";
 
-  useEffect(() => {
-    // Connect to Go WebSocket server
-    const ws = new WebSocket(
-      `ws://localhost:8080/chat-server/ws?hash=${chatHash}&user=${sender}`
-    );
 
-    ws.onopen = () => {
-      console.log(" Connected to WebSocket server");
-    };
 
-    ws.onmessage = (event) => {
-      try {
-        const msg = JSON.parse(event.data);
-        console.log("📩 Received:", msg);
-        setMessages((prev) => [...prev, msg]);
-      } catch {
-        console.log("📩 Received raw:", event.data);
-      }
-    };
+//  Validation function
+function validateUser(Sender: string ): { valid: boolean; error?: string; u1?: string } {
+  const specialCharRegex = /^[a-zA-Z0-9_]+$/;
 
-    ws.onclose = () => {
-      console.log(" Disconnected from server");
-    };
 
-    ws.onerror = (err) => {
-      console.error("WebSocket error:", err);
-    };
+  // Convert to lowercase
+  const u1 = Sender.trim().toLowerCase();
 
-    wsRef.current = ws;
+  if (!specialCharRegex.test(u1)) {
+    return { valid: false, error: "Usernames can only contain letters, numbers, and underscores." };
+  }
 
-    return () => {
-      ws.close();
-    };
-  }, []);
+  return { valid: true, u1};
+}
 
-  const sendMessage = () => {
-    if (!wsRef.current || wsRef.current.readyState !== WebSocket.OPEN) return;
 
-    const outMsg: Message = {
-      hash: chatHash,
-      sender,
-      receiver,
-      message: input,
-    };
 
-    wsRef.current.send(JSON.stringify(outMsg));
-    setMessages((prev) => [...prev, outMsg]); // optimistic update
-    setInput("");
-  };
 
-  return (
-    <div className="p-4">
-      <h2 className="text-xl font-bold mb-2">Chat: {sender} ➝ {receiver}</h2>
-      <div className="border p-2 h-64 overflow-y-auto bg-gray-50 rounded mb-2">
-        {messages.map((m, i) => (
-          <div key={i} className={m.sender === sender ? "text-blue-600" : "text-green-600"}>
-            <b>{m.sender}:</b> {m.message}
-          </div>
-        ))}
-      </div>
+export default function ChatServerRegistration() {
+  const [sender, setSender] = useState('');
+  const [receiver, setReciever] = useState('');
+  const [chatUrl, setChatUrl] = useState('');
+  const [error, setError] = useState('');
+  const [loginForm , setLoginForm] = useState(true) ;
+  const [loginButtonDisabled, setLoginButtonDisabled] = useState(true);
+  const [showMessageSendButton, setMessageSendButton] = useState(false);
+  const [showEndChatButtonDisabled, setEndChatButtonDisabled] = useState(false);
+  const [showLogoutButtonDisabled, setLogoutButtonDisabled] = useState(false);
+  const [showAlert, setShowAlert] = useState(false);
 
-      <div className="flex gap-2">
-        <input
-          className="flex-1 border rounded p-2"
-          value={input}
-          onChange={(e) => setInput(e.target.value)}
-          placeholder="Type a message..."
-        />
-        <button
-          className="bg-blue-500 text-white px-4 py-2 rounded"
-          onClick={sendMessage}
-        >
-          Send
-        </button>
-      </div>
+
+return (
+    <div>
+      <NavBar />
+      <CustomBody>
+
+        <HeadingBar title={"Enter User Name to Login User"} />
+
+            </CustomBody>
+            
+      <Footer />
     </div>
   );
+
 }
